Compilation des classes java

+++ Personne.java
@@ -2,9 +2,9 @@ public class Personne{
 	private String nom;
 	private String prenom;
 	private int age;
-	private float taille;
+	private double taille;
 
-	public Personne(String nom, String prenom, int age, float taille){
+	public Personne(String nom, String prenom, int age, double taille){
 		this.nom = nom;
 		this.prenom = prenom;
 		this.age = age;
